Add: Logic to figure out which operations are descendant related to a specified content base operation.
